Crate airbender_crypto

Re-exports

pub use k256 as rust_k256;

pub use blake2 as blake2_ext;

pub use ark_ec;

pub use ark_ff;

pub use ark_serialize;

Modules

blake2s

bls12_381

bn254

k256

p256

ripemd160

secp256k1

secp256r1

sha3

sha256

Macros

BigInt

Construct a BigInt<N> element from a literal string.

Structs

BigInt

Fp

Represents an element of the prime field F_p, where p == P::MODULUS. This type can represent elements in any field of size at most N * 64 bits.

Enums

BigIntOps

Traits

BigInteger

This defines a BigInteger, a smart wrapper around a sequence of u64 limbs, least-significant limb first.

MiniDigest

Functions

init_lib

parse_u256_be

Parse the byte array as a BE 32-byte BigInt. If length is less than 32 bytes, it will be left-padded (most significant bytes) with zeroes.