airbender_crypto

Module ripemd160

Modules

digest

This crate provides traits which describe functionality of cryptographic hash functions and Message Authentication algorithms.

Structs

Ripemd128Core

Core block-level RIPEMD-128 hasher state.

Ripemd160Core

Core block-level RIPEMD-160 hasher state.

Ripemd256Core

Core block-level RIPEMD-256 hasher state.

Ripemd320Core

Core block-level RIPEMD-320 hasher state.

Traits

Digest

Convenience wrapper trait covering functionality of cryptographic hash functions with fixed output size.

Type Aliases

Ripemd128

RIPEMD-128 hasher.

Ripemd160

RIPEMD-160 hasher.

Ripemd256

RIPEMD-256 hasher.

Ripemd320

RIPEMD-320 hasher.