airbender_crypto

Module sha256

Modules

digest

This crate provides traits which describe functionality of cryptographic hash functions and Message Authentication algorithms.

Structs

Sha256VarCore

Core block-level SHA-256 hasher with variable output size.

Sha512VarCore

Core block-level SHA-512 hasher with variable output size.

Traits

Digest

Convenience wrapper trait covering functionality of cryptographic hash functions with fixed output size.

Type Aliases

Sha224

SHA-224 hasher.

Sha256

SHA-256 hasher.

Sha384

SHA-384 hasher.

Sha512

SHA-512 hasher.

Sha512_224

SHA-512/224 hasher.

Sha512_256

SHA-512/256 hasher.