airbender_crypto

Struct Fp

pub struct Fp<P, const N: usize>(/* private fields */)
where
    P: FpConfig<N>;

Represents an element of the prime field F_p, where p == P::MODULUS. This type can represent elements in any field of size at most N * 64 bits.

Implementations

impl<T, const N: usize> Fp<MontBackend<T, N>, N>
where T: MontConfig<N>,

pub const fn new(element: BigInt<N>) -> Fp<MontBackend<T, N>, N>

Construct a new field element from its underlying BigInt data type.

pub const fn new_unchecked(element: BigInt<N>) -> Fp<MontBackend<T, N>, N>

Construct a new field element from its underlying BigInt data type.

Unlike Self::new, this method does not perform Montgomery reduction. Thus, this method should be used only when constructing an element from an integer that has already been put in Montgomery form.